actuators/requirements.txt
redis
RPi.GPIO
